+++ test/e2e/alcotest/passing/check_basic.ml
@@ -66,6 +66,7 @@ let () =
           id_case (result int unit) "result ok" (Ok 1);
           id_case (result int unit) "result error" (Error ());
           id_case (pair int char) "pair" (1, 'a');
+          id_case (triple int bool string) "triple" (1, true, "a");
         ] );
       ( "negation",
         [